refactor(fe): :zap: re-order import, import with type
