+++ apps/fe/next.config.js
@@ -2,6 +2,9 @@
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { composePlugins, withNx } = require('@nx/next');
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+    enabled: true,
+});
 
 /**
  * @type {import('@nx/next/plugins/with-nx').WithNxOptions}
@@ -58,4 +61,4 @@ const plugins = [
     withNx,
 ];
 
-module.exports = composePlugins(...plugins)(nextConfig);
+module.exports = composePlugins(...plugins)(withBundleAnalyzer(nextConfig));
