+++ apps/fe/src/app/player/[m3u8]/layout.tsx
@@ -1,0 +1,6 @@
+import { Suspense } from 'react';
+import { Skeleton } from 'antd';
+
+export default function NoLayout({ children }: { children: React.ReactNode }) {
+    return <Suspense fallback={<Skeleton />}>{children}</Suspense>;
+}
