+++ apps/fe/src/components/layout/header/index.tsx
@@ -12,11 +12,11 @@ import { stringifyTableParams, useGetIdentity } from '@refinedev/core';
 import { ItemType, MenuItemType } from 'antd/lib/menu/interface';
 import { signOut } from 'next-auth/react';
 
+import { MovieTypeEnum } from 'apps/api/src/app/movies/movie.constant';
+import type { UserDto } from 'apps/api/src/app/users/dtos/user.dto';
+
 import { RouteNameEnum } from '@/constants/route.constant';
 import { getOptimizedImageUrl } from '@/libs/utils/movie.util';
-
-import { MovieTypeEnum } from 'apps/api/src/app/movies/movie.constant';
-import { UserDto } from 'apps/api/src/app/users/dtos/user.dto';
 
 const { Header } = Layout;
 const { useBreakpoint } = Grid;
