+++ apps/fe/src/components/pages/movie/movie-related.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { CrudFilters, stringifyTableParams, useList } from '@refinedev/core';
 
-import { MovieType } from 'apps/api/src/app/movies/movie.type';
+import type { MovieType } from 'apps/api/src/app/movies/movie.type';
 
 import MovieList from '@/components/swiper/movie-list';
 import { MOVIES_LIST_QUERY } from '@/queries/movies';
