+++ apps/fe/src/components/swiper/movie/index.tsx
@@ -1,5 +1,9 @@
-import React, { useEffect, useState } from 'react';
-import type { CSSProperties } from 'react';
+import './movie.css';
+import 'swiper/css';
+import 'swiper/css/effect-fade';
+import 'swiper/css/pagination';
+
+import React, { useEffect, useState, type CSSProperties } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
 import { Typography, Row, Col, Space, Grid, Tag } from 'antd';
@@ -7,12 +11,8 @@ import { CalendarOutlined, EyeOutlined } from '@ant-design/icons';
 import { Swiper, SwiperClass, SwiperSlide } from 'swiper/react';
 import { EffectFade, Pagination, Autoplay } from 'swiper/modules';
 
-import './movie.css';
-import 'swiper/css';
-import 'swiper/css/effect-fade';
-import 'swiper/css/pagination';
-
 import type { MovieResponseDto } from 'apps/api/src/app/movies/dtos';
+
 import { ImageOptimized } from '@/components/image/image-optimized';
 import { MovieQualityTag } from '@/components/tag/movie-quality';
 import { TMDBRating } from '@/components/card/tmdb-rating';
