+++ apps/fe/src/providers/data-provider/restful.ts
@@ -3,9 +3,10 @@
 import { GetListResponse, GetListParams, BaseRecord, DeleteOneParams } from '@refinedev/core';
 import dataProviderSimpleRest from '@refinedev/simple-rest';
 import { AxiosInstance } from 'axios';
-import { handleFilter, handlePagination, handleSort } from '@/libs/utils/data-provider.util';
 
 import type { Movie } from 'apps/api/src/app/movies/movie.schema';
+
+import { handleFilter, handlePagination, handleSort } from '@/libs/utils/data-provider.util';
 
 export const restfulDataProvider = (axios: AxiosInstance) => {
     const apiUrl = `${process.env.NEXT_PUBLIC_API_URL}/api`;
