+++ apps/fe/src/services/movies.ts
@@ -1,4 +1,4 @@
-import { MovieType } from 'apps/api/src/app/movies/movie.type';
+import type { MovieType } from 'apps/api/src/app/movies/movie.type';
 
 export async function getMovieBySlug(slug: string): Promise<MovieType> {
     const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/movies/${slug}`);
